Moved all serial port parsing code into canport.py.
